+++ src/main/java/com/example/minimal_prod_backend/dto/OrdenTrabajoResponse.java
@@ -5,11 +5,11 @@ import java.time.OffsetDateTime;
 public record OrdenTrabajoResponse(
         Long id,
         String referencia,
-        Integer tipoId,
+        Long tipoId,
         String descripcion,
         OffsetDateTime inicioPlanificado,
         OffsetDateTime finPlanificado,
-        Integer estadoId,
+        Long estadoId,
         OffsetDateTime creadoEn
 ) {
 }
